+++ bitcoin/src/blockdata/block.rs
@@ -16,8 +16,7 @@ use super::Weight;
 use crate::blockdata::script;
 use crate::blockdata::transaction::{Transaction, Txid, Wtxid};
 use crate::consensus::{encode, Decodable, Encodable};
-use crate::hash_types::impl_hashencode;
-use crate::internal_macros::impl_consensus_encoding;
+use crate::internal_macros::{impl_consensus_encoding, impl_hashencode};
 use crate::pow::{CompactTarget, Target, Work};
 use crate::prelude::*;
 use crate::{merkle_tree, Network, VarInt};
